Move request hooks off of the Serve trait.
They now exist on an extension trait called RequestHook.
